Move internal.rb inside ruby_lsp
