+++ lib/ruby_lsp/internal.rb
